move MainPod layouting to mainpod
